transferred methods to `DistributedCommmand`, fix codeclimate issues
