+++ src/helpers/help-parser.js
@@ -5,9 +5,10 @@ const path = require('path');
 const fs = require('fs-extra');
 const { exec } = require('child-process-promise');
 const logger = require('./logger');
-const AwsDitributor = require('./distributors/aws-distributor');
+const Dictionary = require('./dictionary');
+const parameters = require('../parameters');
+const AwsDistributor = require('./distributors/aws-distributor');
 const LocalDistributor = require('./distributors/local-distributor');
-const { commandsPath, templates, packageJson } = require('../parameters');
 
 
 class HelpParser {
@@ -16,7 +17,7 @@ class HelpParser {
    * @return {String[]}
    */
   static getCommandsNameList() {
-    return glob.sync('*.js', { cwd: commandsPath }).map(fileName => path.basename(fileName, '.js'));
+    return glob.sync('*.js', { cwd: parameters.commandsPath }).map(fileName => path.basename(fileName, '.js'));
   }
 
   /**
@@ -27,8 +28,8 @@ class HelpParser {
    */
   static getCommandsInstances(list = this.getCommandsNameList()) {
     return list.map(commandName => {
-      const Command = require(path.join(commandsPath, commandName));
-      return new Command(0, logger);
+      const Command = require(path.join(parameters.commandsPath, commandName));
+      return new Command(parameters, logger);
     });
   }
 
@@ -63,7 +64,7 @@ class HelpParser {
    * @param {Boolean} updateBuildDate
    */
   static updateMetadata(updateBuildDate = true) {
-    const packageContent = require(packageJson);
+    const packageContent = require(parameters.packageJson);
     const commands = HelpParser.getCommandsInstances();
     const commandsDescription = HelpParser.getCommandsDescription(commands);
     const { buildDate } = require('../templates/help/metadata');
@@ -76,7 +77,7 @@ class HelpParser {
       commands: commandsDescription
     };
 
-    fs.writeJsonSync(templates.helpMetadata, json, { spaces: 2 });
+    fs.writeJsonSync(parameters.templates.helpMetadata, json, { spaces: 2 });
   }
 
   /**
@@ -86,7 +87,7 @@ class HelpParser {
    * @return {Boolean}
    */
   static hasInvalidOptions(command, args) {
-    const metadata = require(templates.helpMetadata);
+    const metadata = require(parameters.templates.helpMetadata);
     const { options } = metadata.commands.find(it => it.name === command);
 
     return !Object.keys(args).every(arg => options.find(it => it.name === arg || it.shortcut === arg));
@@ -97,7 +98,7 @@ class HelpParser {
    * @return {Promise}
    */
   static updateAWSRegions() {
-    const command = `sh ${path.join(templates.help, 'scripts', 'aws_update.sh')}`;
+    const command = `sh ${path.join(parameters.templates.help, 'scripts', 'aws_update.sh')}`;
 
     return exec(command)
       .then(result => {
@@ -107,7 +108,7 @@ class HelpParser {
           const parsedResult = JSON.parse(stdout);
           const allRegions = [].concat(parsedResult, HelpParser.getPrivateAWSRegions());
 
-          fs.writeJsonSync(path.join(templates.help, 'regions.aws.json'), allRegions, { spaces: 2 });
+          fs.writeJsonSync(path.join(parameters.templates.help, 'regions.aws.json'), allRegions, { spaces: 2 });
         }
       })
       .catch(err => {
